refactor(frontend): unify error pages into responses page and remove error-pages route
